+++ packages/composites/token-allocation/src/dataProvider.ts
@@ -1,6 +1,6 @@
 import { AdapterError, Requester } from '@chainlink/ea-bootstrap'
 import { AdapterResponse, RequestConfig } from '@chainlink/types'
-import { ResponsePayload } from './types'
+import { ResponsePayload, GetPrices } from './types'
 import { Logger } from '@chainlink/ea-bootstrap'
 import { AdapterRequest } from '@chainlink/types'
 
@@ -25,14 +25,16 @@ import { AdapterRequest } from '@chainlink/types'
  * ```
  */
 
-export const getPriceProvider =
-  (source: string, jobRunID: string, apiConfig: RequestConfig) =>
-  async (symbols: string[], quote: string, withMarketCap = false): Promise<ResponsePayload> => {
-    if (source === 'coinpaprika') {
-      return await sendBatchedRequests(source, jobRunID, apiConfig, symbols, quote, withMarketCap)
-    }
-    return await sendIndividualRequests(source, jobRunID, apiConfig, symbols, quote, withMarketCap)
+export const getPriceProvider = (
+  source: string,
+  jobRunID: string,
+  apiConfig: RequestConfig,
+): GetPrices => {
+  if (source === 'coinpaprika') {
+    return sendBatchedRequests(source, jobRunID, apiConfig)
   }
+  return sendIndividualRequests(source, jobRunID, apiConfig)
+}
 
 export interface BatchedAdapterResponse {
   data: {
@@ -40,82 +42,78 @@ export interface BatchedAdapterResponse {
   }
 }
 
-const sendBatchedRequests = async (
-  source: string,
-  jobRunID: string,
-  apiConfig: RequestConfig,
-  symbols: string[],
-  quote: string,
-  withMarketCap = false,
-): Promise<ResponsePayload> => {
-  const sortedSymbols = symbols.sort()
-  const data: AdapterRequest = {
-    id: jobRunID,
-    data: {
-      base: sortedSymbols,
-      quote,
-      endpoint: withMarketCap ? 'marketcap' : 'crypto',
-    },
-  }
-  const responseData = await sendRequestToSource<BatchedAdapterResponse>(source, {
-    ...apiConfig,
-    data,
-  })
-  const tokenPrices = responseData.data.results
-
-  return sortedSymbols.reduce((response, symbol) => {
-    const tokenPrice = tokenPrices.find(
-      (priceResponse) => (priceResponse[0] as AdapterResponse).data.base === symbol,
-    )
-    if (!tokenPrice)
-      throw new AdapterError({
-        jobRunID,
-        statusCode: 500,
-        message: `Cannot find token price result for symbol ${symbol}`,
-      })
-
-    response[symbol] = {
-      quote: {
-        [quote]: { [withMarketCap ? 'marketCap' : 'price']: tokenPrice[1] },
+const sendBatchedRequests =
+  (source: string, jobRunID: string, apiConfig: RequestConfig): GetPrices =>
+  async (symbols, quote, additionalInput, withMarketCap = false): Promise<ResponsePayload> => {
+    const sortedSymbols = symbols.sort()
+    const data: AdapterRequest = {
+      id: jobRunID,
+      data: {
+        ...additionalInput,
+        base: sortedSymbols,
+        quote,
+        endpoint: withMarketCap ? 'marketcap' : 'crypto',
       },
     }
-    return response
-  }, {} as ResponsePayload)
-}
+    const responseData = await sendRequestToSource<BatchedAdapterResponse>(source, {
+      ...apiConfig,
+      data,
+    })
+    const tokenPrices = responseData.data.results
 
-const sendIndividualRequests = async (
-  source: string,
-  jobRunID: string,
-  apiConfig: RequestConfig,
-  symbols: string[],
-  quote: string,
-  withMarketCap = false,
-): Promise<ResponsePayload> => {
-  const results = await Promise.all(
-    symbols.map(async (base) => {
-      const data = {
-        id: jobRunID,
-        data: { base, quote, endpoint: withMarketCap ? 'marketcap' : 'crypto' },
+    return sortedSymbols.reduce((response, symbol) => {
+      const tokenPrice = tokenPrices.find(
+        (priceResponse) => (priceResponse[0] as AdapterResponse).data.base === symbol,
+      )
+      if (!tokenPrice)
+        throw new AdapterError({
+          jobRunID,
+          statusCode: 500,
+          message: `Cannot find token price result for symbol ${symbol}`,
+        })
+
+      response[symbol] = {
+        quote: {
+          [quote]: { [withMarketCap ? 'marketCap' : 'price']: tokenPrice[1] },
+        },
       }
-      const responseData = await sendRequestToSource<AdapterRequest['data']>(source, {
-        ...apiConfig,
-        data,
-      })
-      return responseData.result
-    }),
-  )
-  const payloadEntries = symbols.map((symbol, i) => {
-    const key = symbol
-    const val = {
-      quote: {
-        [quote]: { [withMarketCap ? 'marketCap' : 'price']: results[i] },
-      },
-    }
-    return [key, val]
-  })
+      return response
+    }, {} as ResponsePayload)
+  }
 
-  return Object.fromEntries(payloadEntries)
-}
+const sendIndividualRequests =
+  (source: string, jobRunID: string, apiConfig: RequestConfig): GetPrices =>
+  async (symbols, quote, additionalInput, withMarketCap = false): Promise<ResponsePayload> => {
+    const results = await Promise.all(
+      symbols.map(async (base) => {
+        const data = {
+          id: jobRunID,
+          data: {
+            ...additionalInput,
+            base,
+            quote,
+            endpoint: withMarketCap ? 'marketcap' : 'crypto',
+          },
+        }
+        const responseData = await sendRequestToSource<{ data: { result: number } }>(source, {
+          ...apiConfig,
+          data,
+        })
+        return responseData.data.result
+      }),
+    )
+    const payloadEntries = symbols.map((symbol, i) => {
+      const key = symbol
+      const val = {
+        quote: {
+          [quote]: { [withMarketCap ? 'marketCap' : 'price']: results[i] },
+        },
+      }
+      return [key, val]
+    })
+
+    return Object.fromEntries(payloadEntries)
+  }
 
 const sendRequestToSource = async <T>(source: string, request: AdapterRequest): Promise<T> => {
   try {