+++ packages/sources/ap-election/src/config/index.ts
@@ -7,7 +7,7 @@ export const DEFAULT_ENDPOINT = 'election'
 export const DEFAULT_BASE_URL = 'https://api.ap.org/v2'
 
 export const makeConfig = (prefix?: string): Config => {
-  const config = Requester.getDefaultConfig(prefix)
+  const config = Requester.getDefaultConfig(prefix, true)
   config.api.baseURL = config.api.baseURL || DEFAULT_BASE_URL
   config.defaultEndpoint = DEFAULT_ENDPOINT
   return config