+++ packages/sources/metalsapi/src/endpoint/convert.ts
@@ -20,7 +20,7 @@ export const inputParameters: InputParameters = {
   amount: {
     required: false,
     type: 'number',
-    description: 'The amount fo the `base` currency',
+    description: 'The amount of the `base` currency',
     default: 1,
   },
 }