+++ packages/sources/curve/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/curve-adapter
 
+## 1.2.0
+
+### Minor Changes
+
+- d48deb356: curve: Added description for crypto endpoint
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.1.19
 
 ### Patch Changes