+++ packages/sources/solana-view-function/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/solana-view-function-adapter
 
+## 2.1.6
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 2.1.5
 
 ### Patch Changes