+++ packages/sources/taapi/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/taapi-adapter
 
+## 1.1.20
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.1.19
 
 ### Patch Changes